fix(can): widen txFailCount so CAN_CONF_TX_FAIL_LIMIT is not capped at 255
txFailCount was uint8_t and saturated at UINT8_MAX. Any
CAN_CONF_TX_FAIL_LIMIT above 255 could therefore never be reached, so the
abort recovery would silently never fire - a plausible configuration, e.g.
a 100 ms frame with a ten minute tolerance (6000).

The macro is user facing and documents no upper bound, and the failure is
silent rather than a clamp, so widen the counter instead of constraining
the config.

Free in RAM: the struct is 4 byte aligned and the uint8_t was already
followed by 3 bytes of padding, so sizeof(struct CAN_scheduledMsg) stays
44 either way.

+++ CAN/Inc/can_driver.h
@@ -60,7 +60,7 @@ struct CAN_scheduledMsg
 	uint32_t lastTick;              // time stamp of the last message
 	void (*getData)(uint8_t *data, void *context); // fetches data
 	void *context;                  // user callback context
-	uint8_t txFailCount;            // consecutive failed enqueue attempts, managed by the driver
+	uint32_t txFailCount;           // consecutive failed enqueue attempts, managed by the driver
 };
 
 /**
